+++ src/build.ts
@@ -8,6 +8,7 @@
 const [, , ...aCmd] = process.argv;
 const watch = aCmd.includes('--watch') ?{} :null;
 const web = aCmd.includes('--web') ?{} :null;
+const app = aCmd.includes('--app') ?{} :null;
 
 import {build} from 'vite';
 import dts, {PluginOptions} from 'vite-plugin-dts';
@@ -19,12 +20,13 @@ const output = { // entry chunk assets それぞれの書き出し名の指定
 	assetFileNames: '[name].[ext]',
 };
 const oDts: PluginOptions = {
-	beforeWriteFile: pathOut=> {
-		return {filePath: pathOut.replace('/src/', '/')};
-	},
+	beforeWriteFile: pathOut=> ({
+		filePath: pathOut.replace('/src/', '/'),
+	}),
 };
 
 // === ブラウザ用 ===
+if (! app)
 build({
 	build: {
 		target		: 'esnext',
@@ -38,15 +40,13 @@ build({
 //		minify		: 'terser',
 		reportCompressedSize	: false,
 		watch,
-		rollupOptions: {
-			output,
-		},
+		rollupOptions: {output},
 	},
 	plugins: [dts(oDts)],
 });
-if (! web) {
 
 // === アプリ用 ===
+if (! web)
 build({
 	build: {
 		target		: 'esnext',
@@ -70,6 +70,7 @@ build({
 	plugins: [dts(oDts)],
 });
 
+if (! web && ! app) {
 build({
 	build: {
 		target		: 'esnext',