+++ src/sn/CmnInterface.ts
@@ -63,6 +63,15 @@ export type HSysBaseArg = {
 	crypto	: boolean;
 	dip		: string;
 }
+export type T_SysBaseParams = [
+	hPlg	: HPlugin,
+	arg?	: HSysBaseArg,
+];
+export type T_SysBaseLoadedParams = [
+	hPlg	: HPlugin,
+	arg		: HSysBaseArg,
+];
+
 
 export type SYS_DEC_RET = HTMLImageElement | HTMLVideoElement | ArrayBuffer;
 export interface ISysBase {