+++ src/sn/ConfigBase.ts
@@ -80,8 +80,7 @@ export interface ISysRoots {
 	dec(ext: string, tx: string): Promise<string>;
 	decAB(ab: ArrayBuffer): Promise<HTMLImageElement | HTMLVideoElement | ArrayBuffer>;
 
-	get cur()	: string;
-	get crypto(): boolean;
+	arg: HSysBaseArg;
 	fetch(url: string): Promise<Response>;	// ハッシュ値作成ロード用
 	hash(str: string): string;
 }
@@ -133,8 +132,8 @@ export class ConfigBase implements IConfig {
 	userFnTail		= '';	// 4tst public
 	protected	hPathFn2Exts	: IFn2Path	= {};
 
-	constructor(readonly sys: ISysRoots) {}
-	async load(oCfg: any) {
+	protected	constructor(readonly sys: ISysRoots) {}
+	async load(oCfg: T_CFG) {
 		// this.oCfg = {...this.oCfg, ...oCfg};	// 一階層目でコピーしてしまう
 		this.oCfg.save_ns = oCfg?.save_ns ?? this.oCfg.save_ns;
 
@@ -143,7 +142,7 @@ export class ConfigBase implements IConfig {
 
 		this.oCfg.book = {...this.oCfg.book, ...oCfg.book};
 
-		this.oCfg.log.max_len = oCfg.log?.max_len?.max_len ?? this.oCfg.log.max_len;
+		this.oCfg.log.max_len = oCfg.log?.max_len ?? this.oCfg.log.max_len;
 
 		this.oCfg.init = {...this.oCfg.init, ...oCfg.init};
 
@@ -160,7 +159,7 @@ export class ConfigBase implements IConfig {
 		this.#existsBreakpage = this.matchPath('^breakpage$', SEARCH_PATH_ARG_EXT.SP_GSM).length > 0;
 
 		const hFn2Ext: {[fn: string]: string}	= {};
-		if (! this.sys.crypto) {
+		if (! this.sys.arg.crypto) {
 			for (const [fn0, hExts] of Object.entries(this.hPathFn2Exts)) {
 				for (const ext of Object.keys(hExts)) {
 					if (ext.startsWith(':')) continue;
@@ -275,7 +274,7 @@ export class ConfigBase implements IConfig {
 	addPath(fn: string, h_exts: IExts) {
 		const o: any = {};
 		for (const [ext, v] of Object.entries(h_exts)) {
-			o[ext] = (ext.startsWith(':') ?`` :this.sys.cur) + v;
+			o[ext] = (ext.startsWith(':') ?`` :this.sys.arg.cur) + v;
 		}
 		this.hPathFn2Exts[fn] = o;
 	}