+++ src/sn/SysNode.ts
@@ -13,12 +13,13 @@ import type {IHTag} from './Grammar';
 
 import {Application} from 'pixi.js';
 
+
 export class SysNode extends SysBase {
 	override async loadPath(hPathFn2Exts: IFn2Path, cfg: IConfig) {
 		await super.loadPath(hPathFn2Exts, cfg);
 
 		const fn = this.arg.cur +'path.json';
-		const src = await this.readFileSync(fn);
+		const src = await (await fetch(fn)).text();
 		const oJs = JSON.parse(await this.dec(fn, src));
 		for (const [nm, v] of Object.entries(oJs)) {
 			const h = hPathFn2Exts[nm] = <any>v;