+++ src/sn/SysWeb.ts
@@ -8,7 +8,7 @@
 import {SysBase} from './SysBase';
 import {CmnLib, getDateStr, argChk_Boolean, argChk_Num} from './CmnLib';
 import type {IHTag, ITag} from './Grammar';
-import type {IVariable, IMain, IData4Vari, HPlugin, HSysBaseArg} from './CmnInterface';
+import type {IVariable, IMain, IData4Vari, T_SysBaseParams, T_SysBaseLoadedParams} from './CmnInterface';
 import {Main} from './Main';
 import type {IFn2Path, IConfig} from './ConfigBase';
 
@@ -18,16 +18,16 @@ import 'devtools-detect';
 
 
 export class SysWeb extends SysBase {
-	#path_base	= '';
-	constructor(hPlg = {}, arg = {cur: 'prj/', crypto: false, dip: ''}) {
+	#path_base;
+	constructor(...[hPlg = {}, arg = {cur: 'prj/', crypto: false, dip: ''}]: T_SysBaseParams) {	// DOMContentLoaded は呼び出し側でやる
 		super(hPlg, arg);
 
 		const a = arg.cur.split('/');
 		this.#path_base = (a.length > 2) ? a.slice(0, -2).join('/') +'/' :'';
 
-		this.loaded(hPlg, arg);
+		queueMicrotask(async ()=> this.loaded(hPlg, arg));
 	}
-	protected override async loaded(hPlg: HPlugin, arg: HSysBaseArg) {
+	protected override async loaded(...[hPlg, arg]: T_SysBaseLoadedParams) {
 		await super.loaded(hPlg, arg);
 
 		document.querySelectorAll('[data-prj]').forEach(v=> {
@@ -68,13 +68,7 @@ export class SysWeb extends SysBase {
 		await this.run();
 	}
 	protected	override run = async ()=> {
-		if (this.#main) {
-			const ms_late = 10;	// NOTE: ギャラリーでのえもふり/Live 2D用・魔法数字
-			this.#main.destroy();
-			await new Promise(rs=> setTimeout(rs, ms_late));
-				// clearTimeout()不要と判断
-		}
-
+		if (this.#main) this.#main.destroy();
 		this.#main = new Main(this);
 	}
 	stop() {
@@ -108,7 +102,7 @@ export class SysWeb extends SysBase {
 		hTmp['const.sn.isDebugger'] = (hn === 'localhost' || hn ==='127.0.0.1');
 
 		const ns = this.cfg.getNs();
-		this.flushSub = this.crypto
+		this.flushSub = this.arg.crypto
 		? async ()=> {
 			store.set(ns +'sys_', await this.enc(JSON.stringify(this.data.sys)));
 			store.set(ns +'mark_', await this.enc(JSON.stringify(this.data.mark)));
@@ -131,7 +125,7 @@ export class SysWeb extends SysBase {
 		}
 
 		// データがある場合の処理
-		if (! this.crypto) {
+		if (! this.arg.crypto) {
 			this.data.sys = store.get(ns +'sys');
 			this.data.mark = store.get(ns +'mark');
 			this.data.kidoku = store.get(ns +'kidoku');
@@ -204,12 +198,12 @@ export class SysWeb extends SysBase {
 				'mark': this.data.mark,
 				'kidoku': this.data.kidoku,
 			});
-			const s2 = this.crypto ?await this.enc(s) :s;
+			const s2 = this.arg.crypto ?await this.enc(s) :s;
 			const blob = new Blob([s2], {'type':'text/json'});
 
 			const a = document.createElement('a');
 			a.href = URL.createObjectURL(blob);
-			a.download = (this.crypto ?'' :'no_crypto_')
+			a.download = (this.arg.crypto ?'' :'no_crypto_')
 				+ this.cfg.getNs() + getDateStr('-', '_', '') +'.swpd';
 			a.click();
 
@@ -235,7 +229,7 @@ export class SysWeb extends SysBase {
 		})
 		.then(async blob=> {
 			const s = await blob.text();
-			const o = JSON.parse(this.crypto ?await this.dec('json', s) :s);
+			const o = JSON.parse(this.arg.crypto ?await this.dec('json', s) :s);
 			if (! o.sys || ! o.mark || ! o.kidoku) throw new Error('異常なプレイデータです');
 			if (o.sys[SysBase.VALNM_CFG_NS] !== this.cfg.oCfg.save_ns) {
 				console.error(`別のゲーム【プロジェクト名=${o.sys[SysBase.VALNM_CFG_NS]}】のプレイデータです`);